refactor convert logic
